+++ Events/SRVLaunchedEvent.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using EddiDataDefinitions;
+using System;
 using Utilities;
 
 namespace EddiEvents
@@ -8,7 +9,7 @@ namespace EddiEvents
     {
         public const string NAME = "SRV launched";
         public const string DESCRIPTION = "Triggered when you launch an SRV from your ship";
-        public const string SAMPLE = "{\"timestamp\":\"2016-06-10T14:32:03Z\",\"event\":\"LaunchSRV\",\"Loadout\":\"starter\",\"PlayerControlled\":true}";
+        public const string SAMPLE = @"{ ""timestamp"":""2022-11-24T23:44:25Z"", ""event"":""LaunchSRV"", ""SRVType"":""combat_multicrew_srv_01"", ""SRVType_Localised"":""SRV Scorpion"", ""Loadout"":""default"", ""ID"":53, ""PlayerControlled"":true }";
 
         [PublicAPI("The SRV's loadout")]
         public string loadout { get; private set; }
@@ -19,10 +20,20 @@ namespace EddiEvents
         [PublicAPI("The vehicle ID assigned to the SRV")]
         public int? id { get; private set; }
 
-        public SRVLaunchedEvent(DateTime timestamp, string loadout, bool playercontrolled, int? id) : base(timestamp, NAME)
+        [PublicAPI("The localized SRV type")]
+        public string srvType => vehicleDefinition?.localizedName;
+
+        [PublicAPI("The invariant SRV type")]
+        public string srvTypeInvariant => vehicleDefinition?.invariantName;
+
+        // Not intended to be public facing at this time
+        public VehicleDefinition vehicleDefinition { get; private set; }
+
+        public SRVLaunchedEvent(DateTime timestamp, string loadout, bool playercontrolled, VehicleDefinition vehicleDefinition, int? id) : base(timestamp, NAME)
         {
             this.loadout = loadout;
             this.playercontrolled = playercontrolled;
+            this.vehicleDefinition = vehicleDefinition;
             this.id = id;
         }
     }